Qt: minor tf fixes

+++ modules/qtwidgets/properties/transferfunctioneditor.cpp
@@ -100,7 +100,7 @@ TransferFunctionEditor::TransferFunctionEditor(TransferFunctionProperty* tfPrope
     }
 
     // initialize editor with current tf
-    for (int i = 0; i < transferFunction_->getNumPoints(); ++i) {
+    for (size_t i = 0; i < transferFunction_->getNumPoints(); ++i) {
         onControlPointAdded(transferFunction_->getPoint(i));
     }
