+++ modules/qtwidgets/properties/transferfunctionpropertywidgetqt.cpp
@@ -115,7 +115,7 @@ void TFPushButton::updateFromProperty() {
 
     TransferFunction& transFunc = tfProperty_->get();
     QVector<QGradientStop> gradientStops;
-    for (int i = 0; i < transFunc.getNumPoints(); i++) {
+    for (size_t i = 0; i < transFunc.getNumPoints(); i++) {
         TransferFunctionDataPoint* curPoint = transFunc.getPoint(i);
         vec4 curColor = curPoint->getRGBA();
